refactor, split into seperate files and use a proper base class
