+++ src/marqo/s2_inference/processing/image_utils.py
@@ -3,9 +3,6 @@ import requests
 
 import cv2
 import numpy as np
-import torch
-import torchvision
-
 import onnxruntime
 
 from marqo.s2_inference.s2_inference import get_logger
@@ -13,28 +10,10 @@ from marqo.s2_inference.types import Dict, List, Union, ImageType, Tuple, FloatT
 from marqo.s2_inference.clip_utils import _load_image_from_path
 from marqo.s2_inference.errors import ChunkerError, ChunkerMethodProcessError
 
-from torchvision.models.detection import FasterRCNN_MobileNet_V3_Large_FPN_Weights, fasterrcnn_mobilenet_v3_large_fpn
-from torchvision.models.detection import fasterrcnn_resnet50_fpn_v2, FasterRCNN_ResNet50_FPN_V2_Weights
 from torchvision import transforms
-from torchvision.models.detection import FCOS_ResNet50_FPN_Weights
-
 
 logger = get_logger("image_utils")
 
-
-def get_default_rcnn_params() -> Dict:
-    """sets the default params for a faster-rcnn in pytorch
-
-    Returns:
-        Dict: _description_
-    """
-    return {'box_score_thresh':0.0001, 
-            'box_nms_thresh':0.01, 
-            'rpn_pre_nms_top_n_test':200, 
-            'box_detections_per_img':100,
-            'rpn_post_nms_top_n_test':100, 
-            'min_size':320,
-    }
 
 def get_default_size() -> Tuple:
     """this sets the default image size used for inference for the chunker
@@ -44,130 +23,16 @@ def get_default_size() -> Tuple:
     """
     return (240,240)
 
-def load_pytorch_rcnn():
-    """loads the pytorch faster rcnn model
-
-    Returns:
-        _type_: _description_
-    """
-    weights = FasterRCNN_ResNet50_FPN_V2_Weights.DEFAULT
-    model = fasterrcnn_resnet50_fpn_v2(weights=weights, 
-                         **get_default_rcnn_params()
-                        )
-    # required for detector models otherwise they require targets for inference
-    model.eval()
-
-    # preprocessor lives in the 'weights'
-    preprocess = weights.transforms()
-    
-    return model, preprocess
-
-def load_pytorch_fcos():
-    """this loads the pytorch fcos model
-
-    Returns:
-        _type_: _description_
-    """
-    weights = FCOS_ResNet50_FPN_Weights.DEFAULT
-
-    model = torchvision.models.detection.fcos_resnet50_fpn(weights=weights,
-                        **get_default_rcnn_params())
-    model.eval()
-
-    preprocess = weights.transforms()
-    
-    return model, preprocess
-
 def _PIL_to_opencv(pil_image: ImageType):
 
     if isinstance(pil_image, ImageType):
         return cv2.cvtColor(np.array(pil_image), cv2.COLOR_RGB2BGR)
     raise TypeError(f"expected a PIL image but received {type(pil_image)}")
 
-def demo_postprocess(outputs, img_size, p6=False):
-
-    grids = []
-    expanded_strides = []
-
-    if not p6:
-        strides = [8, 16, 32]
-    else:
-        strides = [8, 16, 32, 64]
-
-    hsizes = [img_size[0] // stride for stride in strides]
-    wsizes = [img_size[1] // stride for stride in strides]
-
-    for hsize, wsize, stride in zip(hsizes, wsizes, strides):
-        xv, yv = np.meshgrid(np.arange(wsize), np.arange(hsize))
-        grid = np.stack((xv, yv), 2).reshape(1, -1, 2)
-        grids.append(grid)
-        shape = grid.shape[:2]
-        expanded_strides.append(np.full((*shape, 1), stride))
-
-    grids = np.concatenate(grids, 1)
-    expanded_strides = np.concatenate(expanded_strides, 1)
-    outputs[..., :2] = (outputs[..., :2] + grids) * expanded_strides
-    outputs[..., 2:4] = np.exp(outputs[..., 2:4]) * expanded_strides
-
-    return outputs
-
-def _infer_yolox(session, preprocess, opencv_image, input_shape):
-
-    img, ratio = preprocess(opencv_image, input_shape)
-
-    ort_inputs = {session.get_inputs()[0].name: img[None, :, :, :]}
-    output = session.run(None, ort_inputs)
-
-    return output, ratio
-
-def _process_yolox(output, ratio, size=(384, 384)):
-
-    predictions = demo_postprocess(output[0], size)[0]
-    boxes = predictions[:, :4]
-    scores = predictions[:, 4:5] 
-
-    boxes_xyxy = np.ones_like(boxes)
-    boxes_xyxy[:, 0] = boxes[:, 0] - boxes[:, 2]/2.
-    boxes_xyxy[:, 1] = boxes[:, 1] - boxes[:, 3]/2.
-    boxes_xyxy[:, 2] = boxes[:, 0] + boxes[:, 2]/2.
-    boxes_xyxy[:, 3] = boxes[:, 1] + boxes[:, 3]/2.
-    boxes_xyxy /= ratio
-
-    return boxes_xyxy, scores
-
-def _filter_yolox(boxes_xyxy, scores):
-
-    # filter
-    inds_f = filter_yolox_boxes(boxes_xyxy)
-    ss = scores.max(-1)
-    boxes_xyxy_filtered = np.array([boxes_xyxy[_i] for _i in inds_f])
-    _s = np.array([ss[_i] for _i in inds_f])
-
-    inds = torchvision.ops.nms(torch.FloatTensor(boxes_xyxy_filtered), torch.FloatTensor(_s), 0.4)
-
-    return boxes_xyxy_filtered[inds]
 
 def _keep_topk(boxes_xyxy, k=10):
     return boxes_xyxy[:k]
 
-def preprocess_yolox(img, input_size, swap=(2, 0, 1)):
-    
-    if len(img.shape) == 3:
-        padded_img = np.ones((input_size[0], input_size[1], 3), dtype=np.uint8) * 114
-    else:
-        padded_img = np.ones(input_size, dtype=np.uint8) * 114
-
-    r = min(input_size[0] / img.shape[0], input_size[1] / img.shape[1])
-    resized_img = cv2.resize(
-        img,
-        (int(img.shape[1] * r), int(img.shape[0] * r)),
-        interpolation=cv2.INTER_LINEAR,
-    ).astype(np.uint8)
-    padded_img[: int(img.shape[0] * r), : int(img.shape[1] * r)] = resized_img
-
-    padded_img = padded_img.transpose(swap)
-    padded_img = np.ascontiguousarray(padded_img, dtype=np.float32)
-    return padded_img, r
 
 def _get_onnx_provider(device: str) -> str:
     """determine where the model should run based on specified device
@@ -185,14 +50,6 @@ def _get_onnx_provider(device: str) -> str:
     logger.info(f"onnx_provider:{fast_onnxprovider}")
     return fast_onnxprovider
 
-def load_yolox_onnx(model_name: str, device: str) -> Tuple[onnxruntime.InferenceSession, preprocess_yolox]:
-
-    fast_onnxprovider = _get_onnx_provider(device)
-
-    session = onnxruntime.InferenceSession(model_name, providers=[fast_onnxprovider])
-    preprocess = preprocess_yolox
-
-    return session, preprocess
 
 def load_rcnn_image(image_name: str, size: Tuple = (320,320)) -> Tuple[ImageType, FloatTensor, Tuple[int, int]]:
     """this is the loading and processing for the input
@@ -418,55 +275,7 @@ def str2bool(string: str) -> bool:
     """
     return string.lower() in ("true", "1", "t")
 
-def load_pretrained_mobilenet():
-    """"
-    loads marqo trained model
-    """
-    model = fasterrcnn_mobilenet_v3_large_fpn(device='cpu', num_classes=1204,
-    box_score_thresh=0.0001, box_nms_thresh=0.01, 
-                            rpn_pre_nms_top_n_test=200, 
-                            box_detections_per_img=100,
-                            rpn_post_nms_top_n_test=100, min_size=320)
 
-    checkpoint_file = 'awesome_mode.pth'
-    checkpoint = torch.load(checkpoint_file, map_location="cpu")
-    weights = FasterRCNN_MobileNet_V3_Large_FPN_Weights.DEFAULT
-    transform = weights.transforms()
-    model.load_state_dict(checkpoint['model'])
-    model.eval()
-    return model, transform
-
-def load_pretrained_mobilenet320():
-    """"
-    loads marqo trained model
-    """
-    model = fasterrcnn_mobilenet_v3_large_fpn(device='cpu', num_classes=1204,
-    box_score_thresh=0.0001, box_nms_thresh=0.01, 
-                            rpn_pre_nms_top_n_test=200, 
-                            box_detections_per_img=100,
-                            rpn_post_nms_top_n_test=100, min_size=320)
-
-    checkpoint_file = 'model_17.pth'
-    checkpoint = torch.load(checkpoint_file, map_location="cpu")
-    weights = FasterRCNN_MobileNet_V3_Large_FPN_Weights.DEFAULT
-    transform = weights.transforms()
-    model.load_state_dict(checkpoint['model'])
-    model.eval()
-    return model, transform
-
-# TODO add YOLOX https://github.com/Megvii-BaseDetection/YOLOX
-# TODO add onnx support https://pytorch.org/vision/0.12/generated/torchvision.models.detection.fasterrcnn_resnet50_fpn.html
-
-def filter_yolox_boxes(boxes, max_aspect_ratio: int = 4, min_area: int = 40*40):
-
-    inds = []
-    for ind,box in enumerate(boxes):
-        box = box.tolist()
-        area = (box[2]-box[0])*(box[3] - box[1])
-        if area > min_area:
-            inds.append(ind)
-    
-    return inds
 
 def replace_small_boxes(boxes, min_area=40*40, new_size=(100,100)):
 
@@ -474,11 +283,11 @@ def replace_small_boxes(boxes, min_area=40*40, new_size=(100,100)):
     for box in boxes:
         area = (box[2]-box[0])*(box[3] - box[1])
         if area < min_area:
-            print(box)
+            #print(box)
             xc = (box[2]-box[0])/2 + box[0]
             yc = (box[3]-box[1])/2 + box[1]
             box = (xc-new_size[0]/2, yc-new_size[1]/2, xc+new_size[0]/2, yc+new_size[1]/2)
-            print(box)
+            #print(box)
         new_boxes.append(box)
     return new_boxes
