+++ src/marqo/s2_inference/processing/yolox_utils.py
@@ -1,0 +1,117 @@
+
+import cv2
+import numpy as np
+import torch
+import torchvision
+
+import onnxruntime
+
+from marqo.s2_inference.s2_inference import get_logger
+from marqo.s2_inference.types import Dict, List, Union, ImageType, Tuple, FloatTensor, ndarray
+
+from marqo.s2_inference.processing.image_utils import _get_onnx_provider
+
+logger = get_logger("image_yolox_utils")
+
+def preprocess_yolox(img, input_size, swap=(2, 0, 1)):
+
+    if len(img.shape) == 3:
+        padded_img = np.ones((input_size[0], input_size[1], 3), dtype=np.uint8) * 114
+    else:
+        padded_img = np.ones(input_size, dtype=np.uint8) * 114
+
+    r = min(input_size[0] / img.shape[0], input_size[1] / img.shape[1])
+    resized_img = cv2.resize(
+        img,
+        (int(img.shape[1] * r), int(img.shape[0] * r)),
+        interpolation=cv2.INTER_LINEAR,
+    ).astype(np.uint8)
+    padded_img[: int(img.shape[0] * r), : int(img.shape[1] * r)] = resized_img
+
+    padded_img = padded_img.transpose(swap)
+    padded_img = np.ascontiguousarray(padded_img, dtype=np.float32)
+    return padded_img, r
+
+def load_yolox_onnx(model_name: str, device: str) -> Tuple[onnxruntime.InferenceSession, preprocess_yolox]:
+
+    fast_onnxprovider = _get_onnx_provider(device)
+
+    session = onnxruntime.InferenceSession(model_name, providers=[fast_onnxprovider])
+    preprocess = preprocess_yolox
+
+    return session, preprocess
+
+def demo_postprocess(outputs, img_size, p6=False):
+
+    grids = []
+    expanded_strides = []
+
+    if not p6:
+        strides = [8, 16, 32]
+    else:
+        strides = [8, 16, 32, 64]
+
+    hsizes = [img_size[0] // stride for stride in strides]
+    wsizes = [img_size[1] // stride for stride in strides]
+
+    for hsize, wsize, stride in zip(hsizes, wsizes, strides):
+        xv, yv = np.meshgrid(np.arange(wsize), np.arange(hsize))
+        grid = np.stack((xv, yv), 2).reshape(1, -1, 2)
+        grids.append(grid)
+        shape = grid.shape[:2]
+        expanded_strides.append(np.full((*shape, 1), stride))
+
+    grids = np.concatenate(grids, 1)
+    expanded_strides = np.concatenate(expanded_strides, 1)
+    outputs[..., :2] = (outputs[..., :2] + grids) * expanded_strides
+    outputs[..., 2:4] = np.exp(outputs[..., 2:4]) * expanded_strides
+
+    return outputs
+
+def _infer_yolox(session, preprocess, opencv_image, input_shape):
+
+    img, ratio = preprocess(opencv_image, input_shape)
+
+    ort_inputs = {session.get_inputs()[0].name: img[None, :, :, :]}
+    output = session.run(None, ort_inputs)
+
+    return output, ratio
+
+def _process_yolox(output, ratio, size=(384, 384)):
+
+    predictions = demo_postprocess(output[0], size)[0]
+    boxes = predictions[:, :4]
+    scores = predictions[:, 4:5] 
+
+    boxes_xyxy = np.ones_like(boxes)
+    boxes_xyxy[:, 0] = boxes[:, 0] - boxes[:, 2]/2.
+    boxes_xyxy[:, 1] = boxes[:, 1] - boxes[:, 3]/2.
+    boxes_xyxy[:, 2] = boxes[:, 0] + boxes[:, 2]/2.
+    boxes_xyxy[:, 3] = boxes[:, 1] + boxes[:, 3]/2.
+    boxes_xyxy /= ratio
+
+    return boxes_xyxy, scores
+
+def filter_yolox_boxes(boxes, max_aspect_ratio: int = 4, min_area: int = 40*40):
+
+    inds = []
+    for ind,box in enumerate(boxes):
+        box = box.tolist()
+        area = (box[2]-box[0])*(box[3] - box[1])
+        if area > min_area:
+            inds.append(ind)
+    
+    return inds
+
+def _filter_yolox(boxes_xyxy, scores):
+
+    # filter
+    inds_f = filter_yolox_boxes(boxes_xyxy)
+    ss = scores.max(-1)
+    boxes_xyxy_filtered = np.array([boxes_xyxy[_i] for _i in inds_f])
+    _s = np.array([ss[_i] for _i in inds_f])
+
+    inds = torchvision.ops.nms(torch.FloatTensor(boxes_xyxy_filtered), torch.FloatTensor(_s), 0.4)
+
+    return boxes_xyxy_filtered[inds]
+
